Survey

Documentation for Survey.

Survey.svydesign — Type

The svydesign object combines a data frame and all the survey design information needed to analyse it.

julia> using Survey; 

julia> load_sample_data(); 

julia> dclus1 = svydesign(id= :dnum, weights= :pw, data = apiclus1, fpc= :fpc) |> print
Survey Design:
data: 183x40 DataFrame
weights: pw
id: dnum
fpc: fpc
nest: false
check_strat: false

Survey.load_sample_data — Method

The Academic Performance Index is computed for all California schools based on standardised testing of students. The data sets contain information for all schools with at least 100 students and for various probability samples of the data.

The load_sample_data() function loads this data into a dataframe called apiclus1

Details about the columns of the dataset can be found here: https://r-survey.r-forge.r-project.org/survey/html/api.html

Survey.svyby — Function

julia> using Survey, Statistics, StatsBase;        

julia> load_sample_data(); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1, fpc=:fpc); 

julia> svyby(:api00, :cname, dclus1, mean)
11×2 DataFrame
 Row │ cname        api00   
     │ String15     Float64 
─────┼──────────────────────
   1 │ Alameda      669.0
   2 │ Fresno       472.0
   3 │ Kern         452.5
   4 │ Los Angeles  647.267
   5 │ Mendocino    623.25
   6 │ Merced       519.25
   7 │ Orange       710.563
   8 │ Plumas       709.556
   9 │ San Diego    659.436
  10 │ San Joaquin  551.189
  11 │ Santa Clara  732.077

Survey.svyby — Function

julia> using Survey, Statistics, StatsBase;   

julia> load_sample_data(); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1, fpc=:fpc);

julia> svyby(:api00, dclus1, mean)
644.1693989071047