Survey

This package is the Julia implementation of the Survey package in R developed by Professor Thomas Lumley.

The need for moving the code to Julia.

At xKDR we processed millions of records from household surveys using the survey package in R. This process took hours of computing time. By implementing the code in Julia, we are able to do the processing in seconds. In this package we have implemented the functions svymean, svyquantile and svysum. We have kept the syntax between the two packages similar so that we can easily move our existing code to the new language.

Documentation for Survey.

Survey.svydesign — Type

The svydesign object combines a data frame and all the survey design information needed to analyse it.

julia> using Survey;

julia> data(api);

julia> dclus1 = svydesign(id= :dnum, weights= :pw, data = apiclus1, fpc= :fpc) |> print
Survey Design:
variables: 183x44 DataFrame
id: dnum
strata: 1, 1, 1 ... (length = 183)
probs: 0.029544719150814778, 0.029544719150814778, 0.029544719150814778 ... (length = 183)
fpc:
    popsize: 757, 757, 757 ... (length = 183)
    sampsize: 183, 183, 183 ... (length = 183)
nest: false
check_strat: true

Survey.svyglm — Type

svyglm(formula, design, dist, link)

The svyglm function can be used to fit glms on svydesign.

julia> using Survey

julia> data(api);

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1);

julia> svyglm(@formula(ell~meals),dclus1,Normal(),IdentityLink())

Survey.svyhist — Method

Histogram plot of a survey design variable.