Survey

This package is the Julia implementation of the Survey package in R developed by Professor Thomas Lumley.

The need for moving the code to Julia.

At xKDR we processed millions of records from household surveys using the survey package in R. This process took hours of computing time. By implementing the code in Julia, we are able to do the processing in seconds. In this package we have implemented the functions svymean, svyquantile and svysum. We have kept the syntax between the two packages similar so that we can easily move our existing code to the new language.

Documentation for Survey.

Survey.svydesign — Type

The svydesign object combines a data frame and all the survey design information needed to analyse it.

julia> using Survey;

julia> data(api);

julia> dclus1 = svydesign(id= :dnum, weights= :pw, data = apiclus1, fpc= :fpc) |> print
Survey Design:
variables: 183x43 DataFrame
id: dnum
strata:
probs: 0.029544719150814778, 0.029544719150814778, 0.029544719150814778 ... (length = 183)
fpc:
    popsize: 757, 757, 757 ... (length = 183)
    sampsize: 183, 183, 183 ... (length = 183)
nest: false
check_strat: true

Survey.svyglm — Type

svyglm(formula, design, dist, link)

The svyglm function can be used to fit glms on svydesign.

julia> using Survey      

julia> data(api); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1); 

julia> svyglm(@formula(ell~meals),dclus1,Normal(),IdentityLink())

Survey.svyby — Function

The svyby function can be used to generate stratified estimates.

julia> using Survey      

julia> data(api); 

julia> dclus1 = svydesign(id=:1, weights=:pw, data = apiclus1); 

julia> svyby(:api00, :cname, dclus1, svymean)
11×3 DataFrame
 Row │ cname        mean     SE
     │ String15     Float64  Float64
─────┼────────────────────────────────
   1 │ Alameda      669.0    16.2135
   2 │ Fresno       472.0     9.85278
   3 │ Kern         452.5    29.5049
   4 │ Los Angeles  647.267  23.5116
   5 │ Mendocino    623.25   24.216
   6 │ Merced       519.25   10.4925
   7 │ Orange       710.562  28.9123
   8 │ Plumas       709.556  13.2174
   9 │ San Diego    659.436  12.2082
  10 │ San Joaquin  551.189  11.578
  11 │ Santa Clara  732.077  12.2291

julia> svyby(:api00, [:cname, :meals], dclus1, svymean);