Survey

This package is the Julia implementation of the Survey package in R developed by Professor Thomas Lumley.

The need for moving the code to Julia.

At xKDR we processed millions of records from household surveys using the survey package in R. This process took hours of computing time. By implementing the code Julia, we are able to do the processing in seconds. The performance comparisions using randomly generated data are in the performance section. We have implemented svymean, svyquantile and svysum function in this package. We have kept the syntax between the two packages similar so that we can easily move our existing code to the new language.

Documentation for Survey.

Survey.svydesign — Type

The svydesign object combines a data frame and all the survey design information needed to analyse it.

julia> using Survey; 

julia> data(api); 

julia> dclus1 = svydesign(id= :dnum, weights= :pw, data = apiclus1, fpc= :fpc) |> print
Survey Design:
data: 183x40 DataFrame
weights: pw
id: dnum
fpc: fpc
nest: false
check_strat: false

Survey.svyby — Function

The svyby function can be used to generate stratified estimates.

julia> using Survey      

julia> data(api); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1, fpc=:fpc); 

julia> svyby(:api00, :cname, dclus1, svymean)
11×2 DataFrame
 Row │ cname        api00   
     │ String15     Float64 
─────┼──────────────────────
   1 │ Alameda      669.0
   2 │ Fresno       472.0
   3 │ Kern         452.5
   4 │ Los Angeles  647.267
   5 │ Mendocino    623.25
   6 │ Merced       519.25
   7 │ Orange       710.563
   8 │ Plumas       709.556
   9 │ San Diego    659.436
  10 │ San Joaquin  551.189
  11 │ Santa Clara  732.077

Survey.svyby — Function

The svyby function can be used to generate stratified estimates. A vector of columns can be used to groupby.

julia> using Survey      

julia> data(api); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1, fpc=:fpc); 

julia> svyby(:api00, [:cname, :meals], dclus1, svymean)

Survey.svyby — Function

The svyby function can be used to generate stratified estimates.

julia> using Survey  

julia> data(api); 

julia> dclus1 = svydesign(id=:dnum, weights=:pw, data = apiclus1, fpc=:fpc);

julia> svyby(:api00, dclus1, svymean)
644.1693989071047